move some funcs to utils.py
