+++ datasets/cache.py
@@ -8,6 +8,7 @@ import xarray as xr
 import numpy as np
 
 from . import registry
+from .utils import get_time_dim, get_lon_lat_dims, numpy_period_string
 from constants import BBOX, COUNTRY_CODE
 
 # paths
@@ -107,53 +108,3 @@ def get_dynamic_function(full_path):
     module = importlib.import_module(module_path)
     
     return getattr(module, function_name)
-
-
-##########
-# Helpers
-
-# TODO: these should go in a separate utils module
-
-def get_time_dim(ds):
-    # get first available time dim
-    time_dim = None
-    for time_name in ['valid_time', 'time']:
-        if hasattr(ds, time_name):
-            time_dim = time_name
-            break
-    if time_dim is None:
-        raise Exception(f'Unable to find time dimension: {ds.coordinates}')
-    
-    return time_dim
-    
-def get_lon_lat_dims(ds):
-    # get first available spatial dim
-    lat_dim = None
-    lon_dim = None
-    for lon_name,lat_name in [('lon','lat'), ('longitude','latitude'), ('x','y')]:
-        if hasattr(ds, lat_name):
-            lat_dim = lat_name
-            lon_dim = lon_name
-            break
-    if lat_dim is None:
-        raise Exception(f'Unable to find space dimension: {ds.coordinates}')
-
-    return lon_dim, lat_dim
-
-def numpy_period_string(t: np.datetime64, period_type: str) -> str:
-    # convert numpy dateime to period string
-    s = np.datetime_as_string(t, unit="s")
-
-    if period_type == "hourly":
-        return s[:13]        # YYYY-MM-DDTHH
-
-    if period_type == "daily":
-        return s[:10]        # YYYY-MM-DD
-
-    if period_type == "monthly":
-        return s[:7]         # YYYY-MM
-
-    if period_type == "yearly":
-        return s[:4]         # YYYY
-
-    raise ValueError(f"Unknown periodType: {period_type}")
